style checks: unignore E713 (always use "not in" for membership test)
This is a straight readability enhancement, with a very small effect on
existing code (three changed instances in two modules). Enforcing this
is mostly to help future contributors who are new to Python learn the
convention (and hopefully see why `thing not in dictvar` is preferable
to `not thing in dictvar`). Existing contributors already use this style
instinctively; the few violations of this rule were in very old code.