xkcd: Operation Cruft Cleanup
I came looking for cruft, but the closest thing I found was outdated
comments referencing "google" even though we haven't used Google for
years. (It was DuckDuckGo for a while, but now it's Bing, because DDG
broke using multiple `site:` queries and never fixed the bug.)

So, now we just talk vaguely about searching the web, instead of Google.

We also have a nicer module-level docstring, which is part of the
cruft-cleaning project as well.