tell: Operation Cruft Cleanup
Use myself (current maintainer) as example, instead of old maintainer
who probably doesn't want to be bothered (or highlighted in case someone
happens to get that help output in a channel where they are).

Sort imports, rename global constant to UPPERCASE, and add a bit of
future-thinking documentation for in-progress work.