pronouns: Operation Cruft Cleanup
On top of updating the comment pointing to the pronoun.is JSON API pull
request we'd like to see merged (the old one was merged into a feature
branch), this also fixes some typos and handles the example commands
properly using the help prefix set in the bot's config.

Oh, and it also removed an unused global logger. You know, actual cruft.