help: be more lenient with expected pastebin URLs
Sometimes it depends on whether you called the API over HTTP or HTTPS.
Sometimes the site just wants to save on HTTPS overhead by returning a
URL with `http://`, hoping nobody will notice/care and correct it.

We always call over HTTPS, but we'll make sure any valid response is
acceptable (with or without the `s`) and enforce HTTPS before output.