tools: use functools.wraps in `deprecated`
The way it worked before, Sphinx would output `**kwargs` as the function
parameters for methods decorated with `deprecated`, instead of the real
signature. Not surprising, I guess, given what the code was doing.

As for why we weren't already using functools.wraps in the decorator
itself, it's been around since the root commit of this repo, ten years
ago, back when this project was still named "Jenni". The functools
module was added in Python 2.5, but maybe this decorator was originally
added before that version came out? We might never know, but let's use
the modern way from now on!