doc: split [core] documentation and CoreSection
The auto-doc of `CoreSection` is now in `config.rst`, with other
auto-doc for `sopel.config`.

However, the user-friendly documentation is kept in `configure.rst`,
where options are grouped. This part is only a draft version, presented
as grouped list of options. This can be used to explain more about
configuration topics:

* how to configure each type of authentication,
* how to configure each type of database,
* how to set logging,

and so on.