coretasks: Major rewrite of MODE handling
It appears (as observed on a real bot in production) that batched MODE
messages would trip up the old code and assign incorrect privilege
bitmasks depending on the ordering and other factors.

This rewrite is a first attempt at fixing that. Notably, it specifically
bails out if something looks too complicated to parse.

Co-authored-by: mal <mal@sec.gd>