tools: undo bad global find/replace match
Replacing "Willie" with "Sopel" literally *everywhere* was a bad choice.
Some references to "Willie" should stay.

Note that this nested formatting doesn't even *work* in Sphinx (inline
formatting can't be nested, for some silly reason), but since this whole
docstring is already overhauled completely on `master` I'll just fix the
typo for 6.6.x and leave the rest to 7.0.