bot: new setup sequence for Sopel
Sopel's setup sequence is now:

* setup logging
* setup plugins
* start Job Scheduler

There are two major change:

* Job Scheduler is now started after setup is complete
* Logging setup is now called from `Sopel.setup`

There is no good reason for the Job Scheduler to be started before the
bot is ready to run, since jobs can be safely added or removed while
it's not running. One may even consider that we should start the Job
Sceduler once the bot is connected only - but that's for another day.