bot: instanciate Trigger and SopelWrapper for each triggered call
When a function is triggered, it is now called with its own instances
of Trigger and SopelWrapper.

The idea is that a function **must not** have any side-effects on other
function's trigger. This makes the bot less prone to race conditions and
other mutability errors (one may argue "mutability horrors").