test: add an --offline option to pytest
Usage:

    $ py.test --offline

This option skips all tests that cannot be run offline, as they are
marked as "online". Either from manual tag:

    import requests

    @pytest.mark.online
    def test_call_google():
        requests.get('https://google.com')

Or from a command example:

    @example('.title http://google.com', '[ Google ] - google.com', online=True)
    def title_command(bot, trigger):
        # ...

This is especially useful for people like me who are always traveling
and can't have access to a reliable Internet connection. This could be
used by Travis to speed-up the test suite, and/or to make it more
reliable.

Existing online examples have been marked as such.