module: fix examples when pytest isn't installed
This broke auto-documenting module commands on the website, at least.
(Good thing I decided to spin up a 'preview' branch of that now, instead
of waiting until after Sopel 7 released to test it out.)

The beauty of `sys.modules` is that its values are module objects. So we
can get the `pytest` module directly from there if it's already imported
(but skip all the test stuff if it's not).

Co-authored-by: Exirel <florian.strzelecki@gmail.com>