config: ListAttribute uses newlines by default
In Sopel 6.x and prior version, multi-line values would be split by
commas:

    option = one, two, three
             and a half, four

would give that:

    ['one', 'two', 'three\nand a half', 'four']

Also, it isn't possible to have a comma inside a value. However, a
fellow Sopelunker, in #1460, added an escape mechanism.

It was a good idea that didn't account for what Python's ConfigParser
could do for us. Indeed, it accepts, as a built-in behavior, to declare
an option on multiple lines. Getting the value would then returns a
string, strip from extra spaces, with values separated by newlines.

Therefore, all I had to do here was to:

* detect if there was a newline character in the value
* if so, split by newlines
* otherwise keep using the comma-separated approach
* remove (sadly) the usage of escape character
* add some tests
* improve docstrings here and there (they were outdated)
* make sure we properly serialize back the values

and voilà!