bot: fix sometimes-undefined variable in flood protection
Since `elapsed` was defined the same way in both places, there's no
reason it can't be defined just once instead, in a higher scope that
allows using it in loop detection also.

The only stumbling block is the fact that the sensible default value for
`elapsed` (0) would trigger logic that shouldn't be triggered. And we
can't use the other sensible default (`None`) because you can't apply
operations like `<` to `NoneType`. So I picked an arbitrary default.