tools.events: add alternate spelling of 462 numeric
I've seen some tutorials/docs around the web use the modern spelling of
this numeric's name. In the interest of making working with Sopel as
easy as possible, I think it's worth including both spellings.

This is akin to HTTP's "Referer", which is spelled incorrectly in the
HTTP standard but is sometimes spelled correctly in related standards
(for example, the DOM spec).