coretasks: additional logic around sending periodic WHO requests
Checks every 30 seconds to see if any channel's last WHO request was
greater than 120 seconds ago, and initiates a request for the channel
with the oldest previous request time.  Does not issue more than one WHO
request every 30 seconds.

The Channel class in target.py was modified to provide a class attribute
to track the last time a WHO request was issued for the channel.