web: fix `__all__` lists & aliases to new package
Draft documentation for the transition to `tools.web` states that old
functions will be available ONLY in `web`, and new functions added
during the life of 7.x will be available ONLY in `tools.web`. The actual
code did not follow this description.

This patch makes sure that even the old, deprecated methods are included
in `__all__`, and removes aliases in `web` to functions added after 6.x
that should appear only in the new `tools.web` namespace.