logger: deprecate sopel.logger.get_logger
It was a nice idea, having a shortcut to get a logger. However, logging
is a python built-in, a de-facto standard since ages, so let's use it
instead.

Beside, the docstring is bonkers. Do not use `get_logger(__name__)`!

Inside a python module `__name__` is its python dotted path. So for
example in `sopel.modules.ip`, using `get_logger(__name__)` would mean
using the logger `sopel.modules.sopel.modules.ip`... which is kind of
dumb.

For external plugins, it's best to provide a good "how to configure
logging for your plugin" instead of using a bad shortcut function that
doesn't do what you would expect.