Prune FUNDING.yml
We've elected not to finish enrolling in Tidelift for now.

GitHub is taking a *very* long time to review applications for the
Sponsors beta/waitlist, and started displaying a big, nasty error if the
file contains any unregistered name(s) that are not Sponsors-enabled.
Let's take out the usernames, too, until we make progress on that front.