clock: make `.t <arg>` more strict
The `.t` command is still a game of "guess what is <arg>?", but now it
doesn't fallback on nick -> channel -> bot -> UTC timezone.

When it detect an error case, it warns the user instead of using a
fallback, and even suggest a procedure to fix the problem when possible.

This is far from being perfect, but we all love the possibility to use
`.t` with a nick, a channel, or a timezone. It's a good tradeoff!