config: Treat '//' as comment start; ignore default prefixes
If we allow '#' to start a comment, the `core.channels` option breaks.
IRC channel names almost always start with '#', so when Sopel saves its
channel list in the new multi-line format, the option will be empty on
next config load.

It's a shame. '#' is the most natural comment prefix in the world for a
Python app like Sopel (because Python itself uses that prefix), but
there's no way around this ConfigParser behavior without completely
replacing the internal file read method. We can't simply tell the parser
not to ignore comments if they're indented; it `strip()`s each line of
whitespace before checking.