config: make ListAttribute values optionally quoted
Used to make `core.channels` option safe in multi-line mode. Channel
names usually start with '#'. ConfigParser interprets any line (after
stripping whitespace) that starts with '#' as a comment and ignores it.
Thus, when Sopel was told to save its config, serializing the list of
channels broke that option and the list would be loaded as empty on the
next restart. Quoting each value fixes that.

Adding this as a feature of ListAttribute itself, rather than doing the
magic elsewhere (like where the channel list is used), made this fix a
lot cleaner. It also means that plugins needing to allow '#' in their
own ListAttribute values can simply toggle this flag instead of having
to implement their own magic.

Credits: @Exirel suggested this solution right before going to bed.
I decided to implement it while he was sleeping, as a surprise. :D