config: quote ListAttribute items when required
In a ListAttribute, a value can start with a #, which is then considered
as a comment by the Config reader. This is bad news, as channel names
usually start with a # (they can start with a & too, which is less
common on freenode).

To prevent this, what I do is:

* quote the item only if necessary
* unquote it if it correspond to a very specific pattern `"#<name>"`

This way, we can have this ListAttribute:

    channels =
        "#sopel"
        &private

Edge case like `"&something"` won't be unquoted, because they don't
match the exact pattern!

(with applied suggestions from code review)

Co-Authored-By: dgw <dgw@technobabbl.es>