loader: consider event callables triggerable too
If a callable has an event set, but no rule, the bot will add a default
'.*' (catch-all) rule when the callable is registered.

This is a developer-friendliness change. Previously, an event-based
trigger also required a catch-all rule decorator ('.*') (but this was
not very well documented).

There should be no behavior change for event handlers that also have
explicit rule decorator(s).

Thanks to @Exirel for nudging me in the right direction when my initial
approach broke everything. :P

Co-Authored-By: Exirel <florian.strzelecki@gmail.com>