loader: is_triggerable rejects jobs
When `clean_callable` runs on a callable, it automatically adds an
`event` attribute: until now, the `is_triggerable` would return `True`
for that callable after that. The problem is that, upon reloading the
bot (after a timeout, a KILL message, etc.), the `event` attribute
would cause job callables to be wrongly counted as triggerable. Thus,
they would be registered with a 'match-any' rule, and throw errors on
every message thereafter.

This bug is the result of a 5+ year old behavior (adding default
`event` to all callables), plus a few-week-old change that would allow
`event` to be considered as triggerable by default, without `rule`.

We caught it thanks to a lucky Sopel Rockstar whose bot running from
`master` happened to receive a KILL message. If that case is rare
enough, it should be noted that this happened when the bot disconnect
(say, because of a network error). Good thing it was catched before
being included in a release!

Co-Authored-By: dgw <dgw@technobabbl.es>