config: add a default `logging_format`
Without this option specified, we get the Python default format, which
is (among other problems) completely devoid of timestamp information.

This format, devised by Exirel while refactoring the logging mechanism,
outputs the timestamp, the package that generated the log line, the log
level of the line, and (finally) the actual message. Taking us from:

    Registered 42 plugins, 0 failed, 0 disabled

to:

    [2019-10-21 12:47:44,272] sopel.bot            INFO     - Registered 42 plugins, 0 failed, 0 disabled

Once this hits the stable release, it will become quite a bit easier to
follow along with bug reports!

Co-Authored-By: Exirel <florian.strzelecki@gmail.com>