find: optimization pass
Removed casts to Identifier on things that already are Identifiers.

Eliminated useless intermediate variables and reassignments in code
working with the bot's memory. Lists are assigned by reference. We can
simply modify the memory in-place using a local alias to the particular
nick's history instead of making a "copy" that isn't actually one and
then assigning it back to the memory field later.