module: fix __all__ nonsense & docs for `event`
Apparently `event` was left out of the __all__ list. I question whether
we should keep it, since it exports literally every name in the module
anyway, but that's for another time.

Right now, the important thing is that `event` is now exported properly,
and the docstring has been updated to account for the fact that a
default `rule` is now set if an event callable does not have one.