bot: fix race-condition in dispatch
The scenario is:

* dispatch is called for a pretrigger
* it looks into the list of callable by priority
* it triggers one of the first callables, which wants to remove a
  callable of the same priority
* because Python must not allow a dict to be modified while it is
  iterated over, Python raises an exception

If we put a lock here, Python won't raise the exception... but we'll get
a deadlock. To prevent that, instead, we copy the list of callables, and
return the ones that are triggered. If any of them wants to change the
callables of the same priority, they won't raise an error, because we
are not iterating anymore on the dict.

Hoo-ray!

(with applied review suggestions from dgw)

Co-Authored-By: dgw <dgw@technobabbl.es>