module: code-style consistency of require_* guards
The privmsg/chanmsg decorators were using a silly trick to get `bot` and
`trigger` out of the passed `*args` by array indexes. It works, and it's
elegant (quite Pythonic, in fact), but most of the decorators explicitly
define those as arguments to the guard function and avoid the hacky way.