reddit: make post info output a separate function taking ID only
Instead of forcing every handler that might want to trigger output of
information about a post (submission) to twist its variables around so
they'll work for a function designed to act on links posted in the chat,
it's much better to have a separate output function that takes just the
ID of a submission, and fetches/displays the details on its own. This
way, no matter what another function is trying to match (Reddit-hosted
images, video links, etc.), all it has to do is pass along the
submission ID -- however it gets that.