trigger: doc overhaul
Added a bunch of stuff for `PreTrigger` (as `.. py:attribute` in Sphinx
because some of the attrs can't be documented inline).

Updated a ton of `Trigger` stuff. Included `:type:` annotations because
new plugin developers will probably find those useful. (Didn't put them
on `PreTrigger` because that's an internal thing, and anyone who's
working in Sopel's core will find it pretty self-explanatory.)