ip: update MaxMind GeoLite2 database fetching
Works with the changes introduced on December 30 (yesterday), which were
announced literally the day before I left on an overseas trip for two
weeks. Obviously, I didn't see the blog post before I left. I wrote this
patch on a motorcoach in the middle of the desert. (Surprisingly, I was
able to validate the fix and push the changes using the bus's Wi-Fi.)

See the blog post from MaxMind for further details on why we have to
make these silly changes (spoiler: it's because California sucks):
https://blog.maxmind.com/2019/12/18/significant-changes-to-accessing-and-using-geolite2-databases/