Tag 7.0.0 RC1 for testing & verification purposes
Releasing "dev0" to TestPyPI last week went well. This time we'll cut a
full-on release candidate and let Travis upload it to the main PyPI,
where it's easier for early adopters to get their hands on the new
version for testing (no need to specify a custom index for pip).

We'll use this RC to verify that the final package installs correctly
through pip, displays correctly on PyPI, and seems to function as
intended in as many situations as our users can throw at it. Having an
essentially final version of 7.0 ready will also help get other projects
ready to go, such as our "official" GitHub plugin (which needs changes
to work properly with the new Sopel release).