Toss notes about non-SQLite DBs + old plugins all over the place
We can make SQLite work pretty much identically to how it did before,
but there are still some things we can't fix. Differing DBAPI behavior
is one of those things, for example placeholders in raw SQL.

Instead of trying to patch over all the little differences and obviating
the purpose of using an ORM in the first place, we'll simply point out
in as many places as possible that old plugins might not work right when
used with a non-SQLite database.

Besides, it's pretty impossible to patch over all the differences unless
we're going to turn Sopel's DB API into an ORM in its own right.