README: Clean up some stuff for PyPI
Since this file is going to be the public face of Sopel 7.0 until the
end of time, some of it needed cleaning up.

We don't want to talk about "modules" anywhere; we call them "plugins"
now, full stop, no exceptions.

Let's also take a more organic approach to sponsorship options.