travis: modernize base build images when possible
We can use py2.7 on Bionic (newest image); it also supports 3.6+. This
is the majority of our required versions, so Bionic is now our default.

The other three Python versions we need to test for the life of Sopel
7.x require special treatment.

Travis only supports py3.3 on the Trusty build image, which is pretty
old by now but obviously still functional.

Our py3.4 & py3.5 builds remain on Xenial, as it's still the newest
build image to support those versions per Travis documentation.