xkcd: suppress link if triggered by a URL match
This is the smallest patch I could devise, at the expense of being kinda
messy. Ideally only one function would need to handle the `commanded`
argument, but due to how the code is already structured it has to get
passed around a bit first.

We could probably do with a small rewrite of this plugin, but not for a
bugfix release. It might happen later if someone feels like it.