bot: fix plugin name in _command_groups
The "_command_groups" attribute is used to generate the documentation
for plugin's command (there is an old piece of code for some
undocumented feature here but that's something else).

The problem is that it tries to derive the plugin's name from the Python
module the callable come from, instead of, you know, the actual plugin's
name.

Since we added the "plugin_name" attribute on all callable back in
8776354b7aae7087f090a0989181400899f6c0d0 (see #1840) we can now use it
safely for the command group.

One-line fix for you all!