bot: allow a callable to be decorated with rule and commands
I modified the loader.clean_callable function to stop adding regex to
the `rule` attribute on a callable that doesn't need to:

* commands,
* action commands,
* nickname commands,

don't need to have a specitic "rule" attribute. Therefore, when the bot
registers a callable, it can register it as an anonymous rule and as a
command, as expected by plugin authors.

Tests have been modified accordingly.