reddit: handle quarantined subreddit lookups
There *is* a property (`subreddit.quarantine`) that says whether the
subreddit object is quarantined. But of course, if the subreddit is in
fact quarantined, attempting to access that (or any other) property
results in a Forbidden error unless the API client has opted in first.

Because that's Good and Helpful API Design™.

So we get to use a wonderful nested try-except that catches the same
exception type twice, just in case the first one was a little white lie.