tld: completely overhaul data fetching from Wikipedia
Replaced those horrid (and ugly) regex patterns with a bare-bones HTML
parser, an in-memory lookup dictionary, and dynamic generation of the
output line depending on what fields are available.

Surely there are some remaining edge cases (though I threw a number of
different TLDs at this and knocked out some already), but this is an
absolutely MASSIVE improvement in maintainability IMO.

Uses the same caching logic as for the IANA TLD list, so any given
Sopel instance will only hit Wikipedia at most once a week or so.