isup: don't assume `.com` any more if dot is missing
The "com" TLD is nowhere near as common as it once was. It's probably
better to just inform the user that they forgot a dot, or tried to look
up a local (unqualified) name that Sopel won't be able to access.

Sopel definitely shouldn't just skip enforcing the presence of a dot,
since LAN names' existence shouldn't be accessible this way. Left a TODO
comment about mitigating this kind of risk.