isup: give specific reason for "down" verdict
Specified timeout values, both to improve response time of the command
and to make the connect/read timeout errors relevant. (By default, all
we'd get is a generic connection error.)

The general `RequestException` is no longer considered as a hard "down".
Catching all of the relevant sub-types means something truly strange has
to happen in order for none of the `except` clauses to match. Sopel's
general exception handling is fine for that, since any occurrences of
that should be reported as possible bugs.