irc: audit for outdated use of "module" vs. "plugin"
I'm doing this because it's easier to drop the change later than skip
over it now and then remember to come back later. There's a good chance
that we don't actually want to change most of this right now.

Oddly, it seems that nothing actually *uses* the `cap_req()` function
provided here. Something to address in a future patch, perhaps.