choose: implement own `strip()`-like function, safe for IRC formatting
Python's built-in `strip()` method for strings considers some control
codes used for IRC formatting to be "whitespace" and removes them if
they appear at the beginning or end of the string. While the built-in
method does take an optional argument listing what characters to strip,
there's no good way to generate our own list to pass in. One can look up
the category of a Unicode character, but there's no way to get a list of
all characters in a category.

So instead, we have a custom function, which also handles adding the
"formatting reset" character if the text contains any formatting. (My
original idea was to write a "parser" of sorts that would insert the
appropriate control code to "close" unmatched formatting, but then I
remembered that the "reset" 0x0F exists and it was Good Enough™.)

Future revisions could gain a little more efficiency on the wire by only
adding the reset character if the text contains an unmatched formatting
code (but that quickly devolves into diminishing returns, so it's likely
not worth doing unless someone gets *really* bored).