plugin: replace intent by ctcp
This is part of #1683: it replaces `sopel.module.intent` by
`sopel.plugin.ctcp`. This does not change how "intents" are managed
internally by Sopel, which should be done later, possibly in Sopel 8 or
maybe Sopel 9.

Also, because CTCP is mostly used for ``ACTION``, this is now the
default: you can use `ctcp`, `ctcp()`, or `ctcp('ACTION')` to achieve
the same result.

Co-authored-by: dgw <dgw@technobabbl.es>