instagram: rework to use `bot.say()`'s `trailing` parameter
This plugin was a bit more complicated to switch over, since the old
order of operations placed the post caption in the middle when parsing
Instagram's full-fat JSON data.

Moving the caption handling to the end (and tweaking it a bit to make
better use of Python language features) makes it possible to let Sopel
handle truncating the caption itself, using the new `bot.say()` param.

oEmbed handling only pulls the author's name and post title, and it's
unlikely to run into over-length issues, but adding `trailing` to it
anyway for safety doesn't cost anything.