wikipedia: use `bot.say()`'s `trailing` parameter for ledes ("snippets")
We sacrifice the ability to know for sure if another Sopel bot's wiki
output was what triggered this call, but gain the ability to include
more of the article text (and test out a hack of the `trailing` param
that will be useful for other plugins with truncatable content that's
not at the very end of the message).

Replaced the `msg_url == trigger` check with a heuristic: Assume that
the triggering user was a Sopel bot if the trigger started with
`[wikipedia] ` and ended with ` | <our wiki page URL>`. Obviously it
won't be a super effective check if two different versions of Sopel with
different Wikipedia output meet, but the old check would also break in
that case. The change seems worthwhile.