instagram: remove "deprecated" (actually removed) oEmbed fallback
As reported in #1996, Facebook "deprecated" the oEmbed API endpoint.
However, instead of a true deprecation where the endpoint simply starts
returning a warning (or even just is marked as deprecated in docs), FB
essentially *removed* the endpoint. It no longer returns valid JSON at
all, let alone the expected result. Sopel receives plain text stating
that the endpoint is "deprecated", with a link for more details.

The replacement API endpoint requires an API key, and this seems like a
good time to consider removing the Instagram plugin to its own package.
I will prepare that change separately, mostly so this fix can be easily
cherry-picked into 7.0.x for maintenance, but also because it could be
enough to wait for removing the plugin entirely until Sopel 8.0.

I'm not personally signed up for Instagram's developer program, so it
wouldn't be fair if I alleged that FB never sent a warning before this
endpoint stopped working. They might have. I'm just taking issue with
the use of the term "deprecated" for something that is *nonfunctional*
rather than merely unmaintained or scheduled for removal.