instagram: remove; will rerelease as standalone plugin if feasible
No promises, but the Instagram plugin will be moved to its own repo
where someone interested could patch it up to accept an API key.

We're just not going to keep it in core long-term, largely because of
our ongoing intent to piece out all nonessential plugins so bug fixes
don't require a whole Sopel release just because an API changed, and so
plugins can be a bit more liberal about extra requirements if their deps
only get installed for users who specifically install that plugin.