irc: prevent bad exit on error
Three important changes in one commit:

The first was to switch from `datetime`'s `now()` to `utcnow()`,
because DST exists and we need to take that into account.

The second one is to switch from `timedelta`'s `seconds` to
`total_seconds()`, because, believe it or not, but a timedelta of
exactly one day means `dt.seconds == 0`.

The last one is a fix for a TODO that has been bugging me since I read
it. Now, when there is a "fatal" error, we decrease the error count
by one per every full 5s that passed since the last error.

Maybe we should do that every time, but I figured it would be an
acceptable tradeoff for now.