irc.backends: call backend.handle_close on timeout and IRC error
Initially, I thought that "handle_close" was called once the socket was
closed. That is not true: it it called when the socket must be closed. I
misunderstood the asynchat/asyncore interface.

Now, on a timeout, the bot will:

* close the connection
* shutdown the timeout scheduler
* shutdown the job scheduler
* call every shutdown routines

I'm glad I dig further into that problem, thanks to the time dedicated
by our users to test our dev version, and for their feedback!