irc: add `finial` parameter to `bot.say()`
Addresses a shortcoming of the `trailing` parameter, which alone is
insufficient to handle some common output patterns used by Sopel's own
core plugins.

For example, the `wikipedia` plugin outputs article snippets in between
quotation marks. Long snippets are truncated and the quote closed by the
`trailing` parameter, but short snippets never invoke `trailing`. They
lose both the closing quote AND the URL linking to the full article.

With this additional parameter, it will be easier for plugins to include
variable-length data in "the middle" of a message and still guarantee
that any important information at the end of the line will be shown.