wikipedia: use new `finial` parameter to `bot.say()` for great justice
Fixes a bug that I never reported in an issue: Missing the URL on short
snippets because the URL was part of `trailing`, which only gets used if
the text is too long for an IRC line.

That bug was the whole inspiration for implementing a new parameter.

The only outstanding problem I see now is that a short first line can
make it appear as if the article ends there:

    [wikipedia] ONT | "ONT or Ont may refer to:" | <link_to_page>

Ideal output would be:

    [wikipedia] ONT | "ONT or Ont may refer to: […]" | <link_to_page>

That's a separate issue, though.