test: first use of `bot.running_triggers` property as intended
The previous `time.sleep()` approach was suggested on IRC, and I just
ran with it because at the time, it was past 2am.

With a fresh mind, though, and a look at the docstrings for both
`bot.running_triggers` and `bot._update_running_triggers()`, I have
realized the true utility of this property.