tests.mocks: `MockIRCServer` joins running triggers' threads by default
With 30 new lines of code (of which only about half is actual code), a
solid 20 lines can be removed from the `isup` tests.

The long-term benefit is that there will be less boilerplate required
when writing tests for plugin callables that are threaded (which is the
default value if the `@plugin.thread` decorator doesn't override it).