Substitute Libera Chat for all public-facing references to freenode
I elected to leave the tests alone, since those could just as well use
`irc.example.com` and work exactly the same way.

Updated the NEWS entry for v6.6.0 just so future release archives and
changelog pages on the website won't point people to an abandoned
channel on freenode. Probably could leave it alone, but it's easy to
update along with everything else.