build: remove EOL Python versions from Travis config
Oldest is now Python 3.6, set to reach end-of-life in December 2021.

This will help us conserve build credits, since Travis CI shifted to
requiring manual grants of credits for open source projects on their
newer infrastructure, leaving behind the straightforward system that
had been in place for many years on their .org platform.

Somewhat related to this: We will be focusing more attention on a move
to GitHub Actions in the coming weeks...