loader: use `inspect` to make callable docs consistent
Our own `trim_docstring` function is OK, but definitely not perfect.
What it returns can be inconsistently indented depending on whether the
docstring passed in is single- or multi-line, or whether a multi-line
docstring starts on line 0 or line 1 (next to or below the """).

Maybe the `inspect` module wasn't so great in the olden days, or maybe
the maintainers of what was then called Willie just weren't aware of its
existence. Whatever the reason, in the modern Python world, there is no
simpler way to get a consistently (de)indented dostring than by using
`inspect.getdoc()`.

May this 7.5-year-old relic of a function (introduced in late 2013 via
15b171c119d62e550f5359770a9b3fa9f59af918) enjoy its brief visit to
deprecation limbo before being removed entirely in Sopel 8.

Co-authored-by: Exirel <florian.strzelecki@gmail.com>