irc: rawlog repr() & no strip(), log decode fail
Print rawlog with repr() so we don't need a hex editor to find problems
with whitespace or non-printable characters.

Don't strip() messages headed for rawlog, which obscures problems with
whitespace.

Add logging and rawlogging of messages that couldn't be decoded, which
were previously silently ignored. New rawlog prefix: `<<!`

Each part of this is a breaking change if any madman is parsing the rawlog,
but I don't think we consider that part of the API.