docs: clarify relationship between `owner` and `owner_account` (again)
We were going to make `core.owner` optional in Sopel 8, but that's not
wise. The setting is used to send important notices to the bot's owner,
and so it has to be set with a proper value even if `core.owner_account`
is being used for verification (e.g. `plugin.require_owner`).