bot, coretasks: add bot.modeparser
Instead of creating an instance of `ModeParser` at each MODE event, the
bot has a `modeparser` attribute with appropriate default values for its
chanmodes, privileges, and param types.

For that to work, when the bot received new ISUPPORT parameters, it
updates the chanmodes and privileges of the mode parser if necessary.

This required to adapt the list of default modes, as `a` and `q` are
more commonly used as privilege rather than modes.

Tests have been updated to manually update the modeparser. This may
require new methods on the test IRC server to streamline an ISUPPORT
configuration, because right now it's a bit too manual.